scad and stl files to build the supports for rpis
basePlate1 is the internal support plate on wich the router, battery and
raspberry pies will screw on. It fits in the case.
rpiPlate3 is the scad file for the raspberry pi plate, made for 3D
printing. The pi screws on this piece (rpiPlate3_2.5beams.stl), and then
on a second piece (rpiPlate3_2.5beams_support.stl) that will be screwed
to the bottom plate.